@@ -2,4 +2,4 @@
 IOCParser modules - Core functionality for extracting and handling IOCs
 
 Author: Marc Rivero | @seifreed
-""" 
+"""
